# cache git config more aggressively