It shaves a couple 100 ms or so from refreshing the status

With `master`:
```
Refreshing magit...done (1.107s, cached 27/53)
Refreshing magit...done (1.107s, cached 27/53)
Refreshing magit...done (1.108s, cached 27/53)
```

With the patch:
```
Refreshing magit...done (0.921s, cached 33/53)
Refreshing magit...done (0.936s, cached 33/53)
Refreshing magit...done (0.951s, cached 33/53)
```