@tarsius 2016-05-30
Yes, let's do that.

I am not a fan of the order of `magit-set`s arguments. To avoid mistakes I would recommend using `(magit-set value "single.variable.string")` when all parts of the variable name are constant. Or maybe we should deprecate `magit-{get,set}` and add `magit-config-{get,set,unset}` like:

```lisp
(defun magit-config-set (&rest args)
  "Set the value of the Git VARIABLE to VALUE.
\n(fn VARIABLE [VARIABLE-PART...] VALUE)"
  (magit-git-string "config"
                    (mapconcat #'identity (butlast args) ".")
                    (car (last args))))

(defun magit-config-unset (variable &rest variable-part)
  (magit-git-string "config" "--unset" (cons variable variable-part)))
```