@npostavs 2016-05-30
While I was writing this code, I had the idea that we could somehow use the `gv` machinery so that `magit-set` could be done with `(setf (magit-get <var>) <value>)`.