@tarsius 2016-05-30
I like your idea better than what I previously suggested.