@npostavs 2016-05-31
Dropping the changes to `magit-set`, it's not really related to this; it just happened to come up when I grepped for `git config` calls. I'll do the `setf` thing later.

---

Oh, by the way, the original timings were on Windows, on GNU/Linux there's not really a difference:

`master`
```
Refreshing magit...done (0.123s, cached 24/52)
Refreshing magit...done (0.124s, cached 24/52)
Refreshing magit...done (0.160s, cached 24/52)
```

with patch
```
Refreshing magit...done (0.161s, cached 29/52)
Refreshing magit...done (0.103s, cached 29/52)
Refreshing magit...done (0.108s, cached 29/52)
```